- update content, - update segmented button
